requirements.txt
pandas
google-generativeai
transformers
scikit-learn
openai
python-dotenv